adds template publish cmd
